requirements.txt
discord.py
pymongo
python-dateutil
